feat(apollo): добавление apollo-клиента
Jira issue: VEGA-567
