+++ .gitignore
@@ -25,3 +25,7 @@ yarn-error.log*
 
 # IDE
 /.idea
+
+# graphql-codegen files
+graphql.schema.json
+__generated__
