+++ codegen.yml
@@ -1,0 +1,32 @@
+overwrite: true
+schema:
+  - http://outsourcing.nat.tepkom.ru:38080/graphql:
+      headers:
+        Authorization: Bearer ${AUTH_TOKEN}
+documents: 'src/**/*.graphql'
+generates:
+  src/__generated__/types.ts:
+    plugins:
+      - 'typescript'
+  ./graphql.schema.json:
+    plugins:
+      - 'introspection'
+  src/:
+    preset: near-operation-file
+    presetConfig:
+      folder: __generated__
+      extension: .tsx
+      baseTypesPath: __generated__/types.ts
+    plugins:
+      - typescript-operations
+      - typescript-react-apollo
+    config:
+      documentMode: 'documentNodeImportFragments'
+      namingConvention: keep
+      nonOptionalTypename: true
+      dedupeOperationSuffix: true
+      omitOperationSuffix: true
+      withComponent: false
+      withHooks: true
+      withHOC: false
+      reactApolloVersion: 3
