+++ package.json
@@ -19,8 +19,11 @@
   },
   "scripts": {
     "build": "cross-env NODE_ENV=production webpack --mode production --progress --config ./webpack.forShellConfig.js",
-    "dev": "cross-env NODE_ENV=development webpack-dev-server --mode development",
-    "dev:shell": "cross-env NODE_ENV=development PORT=9001 webpack-dev-server --mode development --port 9001 --config ./webpack.forShellConfig.js",
+    "dev": "yarn generate:dev:types && cross-env NODE_ENV=development webpack-dev-server --mode development",
+    "dev:shell": "yarn generate:dev:types && cross-env NODE_ENV=development PORT=9001 webpack-dev-server --mode development --port 9001 --config ./webpack.forShellConfig.js",
+    "generate:types": "graphql-codegen --config codegen.yml",
+    "generate:dev:types": "yarn generate:types --require dotenv/config",
+    "generate:dev:types:watch": "yarn generate:dev:types --watch",
     "test": "jest",
     "test:watch": "jest --watch",
     "format": "prettier --write '{src,types,scripts}/**/*.{json,{j,t}s{,x},css,}'",
@@ -33,9 +36,11 @@
     "lint-fix:css": "stylelint 'src/**/*.css' --fix"
   },
   "dependencies": {
+    "@apollo/client": "^3.2.5",
     "@consta/uikit": "^1.10.4",
     "@gpn-prototypes/vega-ui": "^2.13.0",
     "bem-cn": "^3.0.1",
+    "graphql": "^15.4.0",
     "react": "^16.14.0",
     "react-dom": "^16.14.0",
     "react-router-dom": "^5.2.0",
@@ -43,6 +48,12 @@
   },
   "devDependencies": {
     "@gpn-prototypes/frontend-configs": "^1.2.12",
+    "@graphql-codegen/cli": "1.17.10",
+    "@graphql-codegen/introspection": "1.18.0",
+    "@graphql-codegen/near-operation-file-preset": "1.13.3",
+    "@graphql-codegen/typescript": "1.17.10",
+    "@graphql-codegen/typescript-operations": "1.17.8",
+    "@graphql-codegen/typescript-react-apollo": "2.0.7",
     "@testing-library/jest-dom": "^5.11.4",
     "@testing-library/react": "^10.0.3",
     "@types/jest": "^26.0.15",
