--- src/App/App.css
@@ -1,16 +1,0 @@
-.App {
-  &__HeaderMenu {
-    &::after {
-      content: '';
-      width: 1px;
-      height: 18px;
-      background: var(--color-control-bg-border-default);
-    }
-
-    &_disable {
-      &::after {
-        content: none;
-      }
-    }
-  }
-}
