+++ src/App/index.tsx
@@ -1,6 +1,25 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
+import { ApolloClient, HttpLink, InMemoryCache } from '@apollo/client';
 
 import { App } from './App';
 
-ReactDOM.render(<App />, document.getElementById('root'));
+const authToken = localStorage.getItem('auth-token');
+
+const client = new ApolloClient({
+  cache: new InMemoryCache({
+    typePolicies: {
+      Project: {
+        keyFields: ['vid'],
+      },
+    },
+  }),
+  link: new HttpLink({
+    uri: 'http://outsourcing.nat.tepkom.ru:38080/graphql',
+    headers: {
+      Authorization: authToken ? `Bearer ${authToken}` : undefined,
+    },
+  }),
+});
+
+ReactDOM.render(<App graphqlClient={client} />, document.getElementById('root'));
