+++ src/components/BaseHeader/Header.css
@@ -1,0 +1,80 @@
+.VegaHeader {
+  --header-height: 40px;
+  display: grid;
+  border-bottom: 1px solid var(--color-bg-border);
+
+  width: 100%;
+  grid-template-columns: 298px 2px minmax(200px, 1fr);
+  column-gap: 12px;
+
+  height: var(--header-height);
+
+  background: var(--color-bg-secondary);
+
+  &__Delimiter {
+    width: 2px;
+    height: 18px;
+
+    align-self: center;
+
+    background-color: var(--color-bg-border);
+  }
+
+  &__MenuWrap {
+    display: flex;
+    align-items: center;
+
+    min-width: 200px;
+    max-width: 298px;
+  }
+
+  &__MenuTriggerText {
+    overflow: hidden;
+
+    width: 100%;
+
+    white-space: nowrap;
+
+    text-overflow: ellipsis;
+  }
+
+  &__Dropdown {
+    width: 240px;
+
+    background-color: var(--color-bg-default);
+    border: 1px solid rgba(255, 255, 255, 0.2);
+    border-radius: var(--control-radius);
+    box-shadow: 0 4px 16px rgba(0, 32, 51, 0.18), 0 2px 2px rgba(0, 0, 0, 0.04);
+  }
+
+  &__Menu {
+    margin: 0;
+    padding: 12px 0;
+
+    list-style: none;
+  }
+
+  &__MenuItem {
+    padding: 6px 12px;
+
+    &:hover {
+      background-color: var(--color-bg-stripe);
+    }
+  }
+
+  &__MenuDelimiter {
+    margin: 17px 12px;
+
+    height: 1px;
+
+    background-color: var(--color-bg-ghost);
+  }
+
+  &__MenuLink {
+    text-decoration: none;
+  }
+
+  &__Nav {
+    grid-column: 3;
+  }
+}
