+++ webpack.forShellConfig.js
@@ -3,11 +3,12 @@ const singleSpaDefaults = require('webpack-config-single-spa-react-ts');
 const ImportMapPlugin = require('webpack-import-map-plugin');
 const { getAppConfig } = require('./app-config');
 
+// eslint-disable-next-line global-require
 const postCssConfig = { postcssOptions: { ...require('./postcss.config') } };
 
 const { projectName } = getAppConfig();
 
-const externalPackages = ['@gpn-prototypes/vega-ui'];
+const externalPackages = ['@gpn-prototypes/vega-ui', '@apollo/client', 'grapqhl'];
 
 module.exports = (webpackConfigEnv) => {
   const defaultConfig = singleSpaDefaults({
